Update BaseUrl and Record host mapping in controllers
The code updates BaseUrl and Record assignment logic to accommodate different records for different hosts. A `recordWithHost` option has been added, which allows different records to be used based on the host of the incoming request in index.go controller. Configuration file is also modified to include record information for `ipcrystal.com`.