Update app configuration and index template
The application configuration now utilizes a new email address, and changes have been made to the index template. Specifically, the font-family specification was tweaked, some unneeded table row related to 'Connection' was commented out, and a new table row for 'Contact' was added with a mailto link. Footer styling was also slightly adjusted.